+++ src/main/java/jimbot/Launcher.java
@@ -1,0 +1,9 @@
+package jimbot;
+
+import javafx.application.Application;
+
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
