+++ src/main/java/jimbot/Main.java
@@ -1,0 +1,32 @@
+package jimbot;
+
+import java.io.IOException;
+
+import javafx.application.Application;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Scene;
+import javafx.scene.layout.AnchorPane;
+import javafx.stage.Stage;
+import jimbot.ui.MainWindow;
+
+/**
+ * A GUI for Jimbot using FXML.
+ */
+public class Main extends Application {
+
+    private Jimbot jimbot = new Jimbot("data/database.txt");
+
+    @Override
+    public void start(Stage stage) {
+        try {
+            FXMLLoader fxmlLoader = new FXMLLoader(Main.class.getResource("/view/MainWindow.fxml"));
+            AnchorPane ap = fxmlLoader.load();
+            Scene scene = new Scene(ap);
+            stage.setScene(scene);
+            fxmlLoader.<MainWindow>getController().setJimbot(jimbot); // inject the Jimbot instance
+            stage.show();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+}
